pkarr

Struct SignedPacketBuilder

pub struct SignedPacketBuilder { /* private fields */ }

A builder for SignedPacket with many convenient methods, see SignedPacket::builder documentation for examples of how to use this builder.

Implementations

impl SignedPacketBuilder

pub fn record(self, record: ResourceRecord<'_>) -> Self

Insert a ResourceRecord

pub fn rdata(self, name: Name<'_>, rdata: RData<'_>, ttl: u32) -> Self

Insert any type of RData

You can set the name to . to point or the Apex (the public key, of the keypair used in Self::sign)

pub fn a(self, name: Name<'_>, address: Ipv4Addr, ttl: u32) -> Self

Insert an A record.

You can set the name to . to point or the Apex (the public key, of the keypair used in Self::sign)

pub fn aaaa(self, name: Name<'_>, address: Ipv6Addr, ttl: u32) -> Self

Insert an AAAA record.

You can set the name to . to point or the Apex (the public key, of the keypair used in Self::sign)

pub fn address(self, name: Name<'_>, address: IpAddr, ttl: u32) -> Self

Insert an A or AAAA record.

You can set the name to . to point or the Apex (the public key, of the keypair used in Self::sign)

pub fn cname(self, name: Name<'_>, cname: Name<'_>, ttl: u32) -> Self

Insert a CNAME record.

You can set the name to . to point or the Apex (the public key, of the keypair used in Self::sign)

pub fn txt(self, name: Name<'_>, text: TXT<'_>, ttl: u32) -> Self

Insert a TXT record.

You can set the name to . to point or the Apex (the public key, of the keypair used in Self::sign)

pub fn https(self, name: Name<'_>, svcb: SVCB<'_>, ttl: u32) -> Self

Insert an HTTPS record

You can set the name to . to point or the Apex (the public key, of the keypair used in Self::sign)

pub fn svcb(self, name: Name<'_>, svcb: SVCB<'_>, ttl: u32) -> Self

Insert an `SVCB record

You can set the name to . to point or the Apex (the public key, of the keypair used in Self::sign)

pub fn timestamp(self, timestamp: Timestamp) -> Self

Add a custom Timestamp.

pub fn build( self, keypair: &Keypair, ) -> Result<SignedPacket, SignedPacketBuildError>

Alias to Self::sign

pub fn sign( self, keypair: &Keypair, ) -> Result<SignedPacket, SignedPacketBuildError>

Create a Packet from the ResourceRecords inserted so far and sign it with the given Keypair.

Read more about how names will be normalized in SignedPacket::new.